requirements.txt
django==3.0.3
django-bootstrap3
django-crum
